# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Gym OS API — unified client data from TeamUp, Trainerize, and InBody
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export type ClientSummaryEngagementStatus = typeof ClientSummaryEngagementStatus[keyof typeof ClientSummaryEngagementStatus];


export const ClientSummaryEngagementStatus = {
  active: 'active',
  at_risk: 'at_risk',
  disengaged: 'disengaged',
  unknown: 'unknown',
} as const;

export interface ClientSummary {
  id: number;
  name: string;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  photoUrl?: string | null;
  /** @nullable */
  goals?: string | null;
  needsMealPlan: boolean;
  /** @nullable */
  notes?: string | null;
  engagementStatus: ClientSummaryEngagementStatus;
  /** @nullable */
  weeklyAttendanceAvg?: number | null;
  /** @nullable */
  lastAttendanceDate?: string | null;
  /** @nullable */
  lastTrainingDate?: string | null;
  /** @nullable */
  workoutCompliancePct?: number | null;
  /** @nullable */
  latestScanDate?: string | null;
  /** @nullable */
  latestWeight?: number | null;
  /** @nullable */
  latestBodyFatPct?: number | null;
  /** @nullable */
  latestMuscleMass?: number | null;
  /** @nullable */
  lastSyncedAt?: string | null;
  /** @nullable */
  teamupId?: string | null;
  /** @nullable */
  trainerizeId?: string | null;
  /** @nullable */
  inbodyId?: string | null;
}

export type ClientDetailEngagementStatus = typeof ClientDetailEngagementStatus[keyof typeof ClientDetailEngagementStatus];


export const ClientDetailEngagementStatus = {
  active: 'active',
  at_risk: 'at_risk',
  disengaged: 'disengaged',
  unknown: 'unknown',
} as const;

export interface InBodyScan {
  id: number;
  clientId: number;
  scannedAt: string;
  /** @nullable */
  weightKg?: number | null;
  /** @nullable */
  bodyFatPct?: number | null;
  /** @nullable */
  muscleMassKg?: number | null;
  /** @nullable */
  bmr?: number | null;
  /** @nullable */
  visceralFatLevel?: number | null;
  /** @nullable */
  bmi?: number | null;
  /** @nullable */
  totalBodyWaterKg?: number | null;
}

export interface AttendanceRecord {
  id: number;
  clientId: number;
  date: string;
  /** @nullable */
  className?: string | null;
}

export interface ClientDetail {
  id: number;
  name: string;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  photoUrl?: string | null;
  /** @nullable */
  goals?: string | null;
  needsMealPlan: boolean;
  /** @nullable */
  notes?: string | null;
  engagementStatus: ClientDetailEngagementStatus;
  /** @nullable */
  weeklyAttendanceAvg?: number | null;
  /** @nullable */
  lastAttendanceDate?: string | null;
  /** @nullable */
  lastTrainingDate?: string | null;
  /** @nullable */
  workoutCompliancePct?: number | null;
  /** @nullable */
  latestScanDate?: string | null;
  /** @nullable */
  latestWeight?: number | null;
  /** @nullable */
  latestBodyFatPct?: number | null;
  /** @nullable */
  latestMuscleMass?: number | null;
  /** @nullable */
  lastSyncedAt?: string | null;
  /** @nullable */
  teamupId?: string | null;
  /** @nullable */
  trainerizeId?: string | null;
  /** @nullable */
  inbodyId?: string | null;
  recentScans: InBodyScan[];
  recentAttendance: AttendanceRecord[];
}

export interface ClientUpdate {
  /** @nullable */
  goals?: string | null;
  needsMealPlan?: boolean;
  /** @nullable */
  notes?: string | null;
}

export interface SyncResult {
  success: boolean;
  clientsUpdated: number;
  attendanceRecordsAdded: number;
  trainingSessionsAdded: number;
  scansAdded: number;
  syncedAt: string;
  errors: string[];
  configuredSources: string[];
  missingSources: string[];
}

export interface DashboardStats {
  totalClients: number;
  activeClients: number;
  atRiskClients: number;
  disengagedClients: number;
  needsMealPlanCount: number;
  overdueInBodyCount: number;
  /** @nullable */
  lastSyncedAt?: string | null;
}

export type ListClientsParams = {
search?: string;
engagementStatus?: ListClientsEngagementStatus;
needsMealPlan?: boolean;
};

export type ListClientsEngagementStatus = typeof ListClientsEngagementStatus[keyof typeof ListClientsEngagementStatus];


export const ListClientsEngagementStatus = {
  active: 'active',
  at_risk: 'at_risk',
  disengaged: 'disengaged',
} as const;