# lib/api-zod/src/generated/types/attendanceRecord.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Gym OS API — unified client data from TeamUp, Trainerize, and InBody
 * OpenAPI spec version: 0.1.0
 */

export interface AttendanceRecord {
  id: number;
  clientId: number;
  date: string;
  /** @nullable */
  className?: string | null;
}