# lib/api-zod/src/generated/types/clientSummary.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Gym OS API — unified client data from TeamUp, Trainerize, and InBody
 * OpenAPI spec version: 0.1.0
 */
import type { ClientSummaryEngagementStatus } from './clientSummaryEngagementStatus';

export interface ClientSummary {
  id: number;
  name: string;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  photoUrl?: string | null;
  /** @nullable */
  goals?: string | null;
  needsMealPlan: boolean;
  /** @nullable */
  notes?: string | null;
  engagementStatus: ClientSummaryEngagementStatus;
  /** @nullable */
  weeklyAttendanceAvg?: number | null;
  /** @nullable */
  lastAttendanceDate?: string | null;
  /** @nullable */
  lastTrainingDate?: string | null;
  /** @nullable */
  workoutCompliancePct?: number | null;
  /** @nullable */
  latestScanDate?: string | null;
  /** @nullable */
  latestWeight?: number | null;
  /** @nullable */
  latestBodyFatPct?: number | null;
  /** @nullable */
  latestMuscleMass?: number | null;
  /** @nullable */
  lastSyncedAt?: string | null;
  /** @nullable */
  teamupId?: string | null;
  /** @nullable */
  trainerizeId?: string | null;
  /** @nullable */
  inbodyId?: string | null;
}