# lib/api-zod/src/generated/types/dashboardStats.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Gym OS API — unified client data from TeamUp, Trainerize, and InBody
 * OpenAPI spec version: 0.1.0
 */

export interface DashboardStats {
  totalClients: number;
  activeClients: number;
  atRiskClients: number;
  disengagedClients: number;
  needsMealPlanCount: number;
  overdueInBodyCount: number;
  /** @nullable */
  lastSyncedAt?: string | null;
}