# lib/api-zod/src/generated/types/inBodyScan.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Gym OS API — unified client data from TeamUp, Trainerize, and InBody
 * OpenAPI spec version: 0.1.0
 */

export interface InBodyScan {
  id: number;
  clientId: number;
  scannedAt: string;
  /** @nullable */
  weightKg?: number | null;
  /** @nullable */
  bodyFatPct?: number | null;
  /** @nullable */
  muscleMassKg?: number | null;
  /** @nullable */
  bmr?: number | null;
  /** @nullable */
  visceralFatLevel?: number | null;
  /** @nullable */
  bmi?: number | null;
  /** @nullable */
  totalBodyWaterKg?: number | null;
}